refactor/6000: utils관련 파일 디렉토리 이동
