+++ src/screens/SignIn.js
@@ -9,12 +9,12 @@ import {
   StyleSheet,
 } from 'react-native';
 import React, { useState, useEffect, useRef } from 'react';
-import { signIn } from '../../firebase';
+import { signIn } from '../utils/firebase';
 import { onAuthStateChanged } from 'firebase/auth';
-import { auth } from '../../firebase';
+import { auth } from '../utils/firebase';
 import { UserAuth } from '../context/AuthContext';
 import { theme } from '../../theme';
-import { validateEmail, removeWhitespace } from '../../utils';
+import { validateEmail, removeWhitespace } from '../utils/utils';
 import ErrorMessage from '../components/ErrorMessage';
 
 const colors = theme.colors;
